+++ frontend/src/main/resources/webapp/src/app/main/books/cards/book-card-medium/book-card-medium.component.ts
@@ -1,5 +1,5 @@
 import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
-import {Book} from "../../book";
+import {Book} from "../../../../shared/domain/book";
 import {LoginService} from "ontimize-web-ngx";
 
 @Component({