+++ frontend/src/main/resources/webapp/src/app/shared/domain/user.ts
@@ -1,0 +1,14 @@
+export interface User {
+  user_: string;
+  password?: string;
+  name?: string;
+  surname?: string;
+  email: string;
+  nif?: string;
+  birthdate?: Date;
+  avatar?: string;
+  userblocked?: Date;
+  lastpasswordupdate?: Date;
+  firstlogin?: boolean;
+  enabled?: boolean;
+}